+++ src/acquirium/__init__.py
@@ -1,9 +1,8 @@
 from acquirium.Client.acquirium import Acquirium
-from acquirium.Apps.base import App, Output
+from acquirium.Materialization import App, OutputSpec, align, output
 from acquirium.Client.explore.core import Query
 from acquirium.Client.query import Q
 from acquirium.Client.data_object import DataObject
-from acquirium.internals.models import AppContext
 from acquirium.Drivers.Driver import (
     Driver,
     DriverBufferFull,